backend/logging_monitoring/exporters/prometheus.py
"""
Prometheus metrics exporter

Note: This is a mock implementation for development.
In production, use prometheus_client library.
"""

from typing import Dict, List, Any
from datetime import datetime


class PrometheusExporter:
    """
    Prometheus metrics exporter
    
    Mock implementation for development. In production, integrate with
    prometheus_client library:
    
    from prometheus_client import Counter, Histogram, Gauge, generate_latest
    """
    
    def __init__(self, port: int = 9090, path: str = "/metrics"):
        self.port = port
        self.path = path
        self.metrics: Dict[str, Any] = {}
        self.enabled = False
    
    def enable(self) -> None:
        """Enable Prometheus exporter"""
        self.enabled = True
        print(f"[Prometheus Exporter] Enabled on port {self.port} at {self.path}")
        print("[Prometheus Exporter] Note: This is a mock implementation")
        print("[Prometheus Exporter] For production, install: pip install prometheus-client")
    
    def register_counter(self, name: str, description: str, labels: List[str] = None) -> None:
        """Register a counter metric"""
        if labels is None:
            labels = []
        
        self.metrics[name] = {
            "type": "counter",
            "description": description,
            "labels": labels,
            "value": 0,
        }
    
    def register_gauge(self, name: str, description: str, labels: List[str] = None) -> None:
        """Register a gauge metric"""
        if labels is None:
            labels = []
        
        self.metrics[name] = {
            "type": "gauge",
            "description": description,
            "labels": labels,
            "value": 0,
        }
    
    def register_histogram(self, name: str, description: str, labels: List[str] = None) -> None:
        """Register a histogram metric"""
        if labels is None:
            labels = []
        
        self.metrics[name] = {
            "type": "histogram",
            "description": description,
            "labels": labels,
            "values": [],
        }
    
    def increment(self, name: str, value: float = 1.0, labels: Dict[str, str] = None) -> None:
        """Increment a counter"""
        if name in self.metrics and self.metrics[name]["type"] == "counter":
            self.metrics[name]["value"] += value
    
    def set_gauge(self, name: str, value: float, labels: Dict[str, str] = None) -> None:
        """Set a gauge value"""
        if name in self.metrics and self.metrics[name]["type"] == "gauge":
            self.metrics[name]["value"] = value
    
    def observe(self, name: str, value: float, labels: Dict[str, str] = None) -> None:
        """Observe a value for histogram"""
        if name in self.metrics and self.metrics[name]["type"] == "histogram":
            self.metrics[name]["values"].append(value)
    
    def generate_metrics(self) -> str:
        """
        Generate Prometheus metrics format
        
        Returns:
            Metrics in Prometheus text format
        """
        lines = []
        
        for name, metric in self.metrics.items():
            # Add HELP
            lines.append(f"# HELP {name} {metric['description']}")
            
            # Add TYPE
            lines.append(f"# TYPE {name} {metric['type']}")
            
            # Add metric value
            if metric["type"] in ["counter", "gauge"]:
                lines.append(f"{name} {metric['value']}")
            elif metric["type"] == "histogram":
                # Simplified histogram output
                values = metric["values"]
                if values:
                    lines.append(f"{name}_count {len(values)}")
                    lines.append(f"{name}_sum {sum(values)}")
        
        return "\n".join(lines)


# Mock implementation notice
def get_prometheus_exporter(enabled: bool = False) -> PrometheusExporter:
    """
    Get Prometheus exporter instance
    
    For production use, install prometheus_client:
    pip install prometheus-client
    
    Then use:
    from prometheus_client import Counter, Histogram, Gauge, start_http_server
    """
    exporter = PrometheusExporter()
    
    if enabled:
        exporter.enable()
        print("\n" + "="*60)
        print("PROMETHEUS INTEGRATION SETUP")
        print("="*60)
        print("\nFor production, install prometheus-client:")
        print("  pip install prometheus-client")
        print("\nExample usage:")
        print("  from prometheus_client import Counter, start_http_server")
        print("  ")
        print("  # Start metrics server")
        print("  start_http_server(9090)")
        print("  ")
        print("  # Create metrics")
        print("  requests_total = Counter('requests_total', 'Total requests')")
        print("  requests_total.inc()")
        print("="*60 + "\n")
    
    return exporter
