backend/requirements.txt
# Backend Python Dependencies
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
watchdog>=3.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
aiohttp>=3.8.0
pytest-cov>=4.1.0
